Unusual behaviour more clear, & move 1 sentence
This function's unusual behaviour tripped me up. I ended up posting a question to StackOverflow because I didn't expect the unusual behaviour (see 'https://stackoverflow.com/questions/55659056/why-is-vbas-vartype-function-saying-this-com-object-is-a-string-object-is-ins').

I believe the function name, along with what was previously specified in the 'summary' section of the documentation for the function, probably contributed to my not realising that the function had unusual behaviour.

I've made changes to the documentation to make it more clear that the function has unusual behaviour. The 'summary' section before, was not strictly correct (in my opinion), as it gave the impression that this function didn't have any unusual behaviour.

I've also moved one of the sentences in the Remarks section as it seems that the sentence was put into the wrong paragraph.