Fixed code example
This `GoTo` jump is harmful and suggests that it's ok to resume from an error-handling subroutine with `GoTo` - it's not. Run-time is still in an error state when execution enters `EndRoutine`, contrary to `Resume EndRoutine`, which correctly clears the error state.
Removed the redundant `Set ... = Nothing` assignments; fully-qualified all Outlook types, to make the dependency explicit (since people are copy-pasting this and then making it late-bound); tightened vertical whitespace, moved variable declarations closer to their usage, and removed the harmful and outdated Systems Hungarian Notation that is heavily discouraged everywhere.