Added content useful to 'understanding variants'
Because the page title is 'Understanding Variants', I believe the page should be more complete in respect of such understanding. I have made changes to this effect.

I added a reference to the page on the variant data type in the 'See also' section. I believe this should be because even though there is a link higher up on this 'understanding' page, to this variant-data-type page, you can't tell whether the higher-up link goes to the glossary definition or some other page. The linked-to page is most definitely useful reading following on from this 'understanding' page and is not simply a short paragraph defining the type. Because of this, I feel that there should be a link in the 'See also' section.