Corrections & improvements; esp. re. UDTs
User-defined types can be used with TypeName, which is shown in the amended example on this page. Have made changes for this.

Example has been increased to demonstrate how to obtain more values using TypeName, especially those values that are not so regular.

Added reference to VarType function in 'See also' section. VarType has strong similarities to TypeName.

Other changes are for slight mistakes.