VarType or TypeName alone can't determine sub-type
To see that VarType doesn't always determine the sub-type, consult the documentation for VarType at: https://docs.microsoft.com/en-us/office/vba/language/reference/user-interface-help/vartype-function .

You can also see this inability of VarType by running the following code:

    Set v = Excel.Application
    MsgBox "Set v = Excel.Application" & vbCrLf & vbCrLf & _
            vbTab & "TypeName(v) : " & vbTab & """" & TypeName(v) & """" & vbCrLf & _
            vbTab & "VarType(v) : " & vbTab & VarType(v) & vbCrLf & vbCrLf & _
            "VBA.VbVarType.vbString: " & VBA.VbVarType.vbString & vbCrLf & vbCrLf & _
            vbTab & "IsObject(v): " & vbTab & vbTab & IsObject(v) & vbCrLf & _
            vbTab & "IsObject(Excel.Application): " & vbTab & IsObject(Excel.Application) & vbCrLf & vbCrLf & _
            "Excel.Application.Caption: " & vbTab & """" & Excel.Application.Caption & """" & vbCrLf & _
            "v.Caption: " & vbTab & vbTab & """" & v.Caption & """"

The changes I have made indicate to users that the VarType function should be used in conjunction with the IsObject function, to '...determine how the data in a Variant is treated...'.

The TypeName function is incapable of determining the sub-type of a variant in at least the following scenarios:
 - The return value indicates either an object or a UDT but you can't figure out which it is, possibly because:
	+ The value returned belongs to both a class and a UDT, or
	+ You don't recognise the value because you have limited information regarding class names and / or UDT names.
 - An object or UDT has a name with a value that TypeName returns when given a data type that is neither a UDT nor an object. For example, the mscorlib library has a 'String' class - 'String' is also the value returned for the primitive string data type.
 
Because of TypeName's inability, my suggested change includes removing the text where TypeName is stated as being able to determine how the data in a Variant is treated unassisted. Instead, I've added information about how TypeName can increase the clarity of initial determination.