UDTs from lib. refs; Variant def. not right
"User-defined type" glossary definition
=========================
As far as I can tell, types contained in project library references are also considered user-defined types (UDTs). This seems to be the case because:
 - VarType says such types are UDTs.
 - Such types seem to be generally interchangeable with UDTs declared using the Type statement.
 - VBA's primitive types don't appear to be contained in such project library references.

I used the OleView.exe tool to examine the mscorlib.tlb type library (the type library for a core .NET library). The Guid type defined in the type library, appears to be defined as follows in the OleView.exe-tool output:

    typedef [uuid(9C5923E9-DE52-33EA-88DE-7EBC8633B9CC), version(1.0)    ,
      custom(0F21F359-AB84-41E8-9A78-36D110E6D2F9, "System.Guid")    
     ]
    struct tagGuid { long _a; short _b; short _c; unsigned char _d; unsigned char _e; unsigned char _f; unsigned char _g; unsigned char _h;
                              unsigned char _i; unsigned char _j; unsigned char _k;
    } Guid;

You'll notice that no 'Type' statement is used in this definition. Because of this, I believe that it is entirely possible for such types not to be defined using a 'Type' statement as the type library can be developed using another programming language.

Because of the above, I believe the previous glossary definition for UDTs very likely needs to also include mention of such types (types available through the use of external project references.)

-------------------------------------------------------------------------------------------------------------------------------------------------

"Variant type" glossary definition
=====================
Previous definition didn't mention that the variant type can store object types. It also stated that VarType defines how data in variants are treated. Don't think 'define' is the right word unless the function is used 'behind-the-scenes' by the variant data type. The VarType function does not always indicate the sub-type. To see this is the case, consult the VarType documentation at 'https://docs.microsoft.com/en-us/office/vba/language/reference/user-interface-help/vartype-function' or run the following code:

    Set v = Excel.Application
    MsgBox "Set v = Excel.Application" & vbCrLf & vbCrLf & _
            vbTab & "TypeName(v) : " & vbTab & """" & TypeName(v) & """" & vbCrLf & _
            vbTab & "VarType(v) : " & vbTab & VarType(v) & vbCrLf & vbCrLf & _
            "VBA.VbVarType.vbString: " & VBA.VbVarType.vbString & vbCrLf & vbCrLf & _
            vbTab & "IsObject(v): " & vbTab & vbTab & IsObject(v) & vbCrLf & _
            vbTab & "IsObject(Excel.Application): " & vbTab & IsObject(Excel.Application) & vbCrLf & vbCrLf & _
            "Excel.Application.Caption: " & vbTab & """" & Excel.Application.Caption & """" & vbCrLf & _
            "v.Caption: " & vbTab & vbTab & """" & v.Caption & """"

Because of this, I have changed the reference to VarType to a reference to VarType used in conjunction with IsObject.

Also made a minor grammatical correction.