Fixes & improvements; esp. of TypeOf & condition
Appears to have been quite a few mistakes in the 'condition' section of the table. Can someone please double-check these corrections?

I have no idea why "TypeOf...Is.." is documented on this page. VB .NET documents it as being an operator in the VB.NET (https://docs.microsoft.com/en-us/dotnet/visual-basic/language-reference/operators/typeof-operator) documentation, and my own thinking is that it is an operator in VBA too. If this is right, it should be documented in the Operators section of the documentation & not here.

As an intermediate fix, I've tried to correct & improve the 'TypeOf...Is...' documentation on this page. An ultimate goal, as far as I can tell, should be to move this documentation elsewhere (into hopefully the right place.)

Other changes are minor corrections & minor improvements.