Fixes; things missing + some deletions
As far as I can tell, types contained in project library references are also considered user-defined types (UDTs). This seems to be the case because:
 - VarType says such types are UDTs.
 - Such types seem to be generally interchangeable with UDTs declared using the Type statement.
 - VBA's primitive types don't appear to be contained in such project library references.

I used the OleView.exe tool to examine the mscorlib.tlb type library (the type library for a core .NET library). The Guid type defined in the type library, appears to be defined as follows in the OleView.exe-tool output:

    typedef [uuid(9C5923E9-DE52-33EA-88DE-7EBC8633B9CC), version(1.0)    ,
      custom(0F21F359-AB84-41E8-9A78-36D110E6D2F9, "System.Guid")    
     ]
    struct tagGuid { long _a; short _b; short _c; unsigned char _d; unsigned char _e; unsigned char _f; unsigned char _g; unsigned char _h;
                              unsigned char _i; unsigned char _j; unsigned char _k;
    } Guid;

You'll notice that no 'Type' statement is used in this definition. Because of this, I believe that it is entirely possible for such types not to be defined using a 'Type' statement as the type library can be developed using another programming language.

Because of the above, I believe the previous glossary definition for UDTs very likely needs to also include mention of such types (types available through the use of external project references.) The 'User-defined (using Type or other means)' text in the present commit has been written by myself because of these things.

-----------------------------------------------------

The other changes are self-explanatory.

-----------------------------------------------------

The 'Returns for CStr' section should probably be removed in a future commit as having the 'CStr' documentation in two places makes it hard to maintain and adds confusion to users.