Intrinsic & non-intrinsic fixed + improvements
The former intrinsic-data-type section wasn't right. Have now tried to document intrinsic data types & non-intrinsic data types separately in a clear & complete way.

Have also made some relatively minor improvements / fixes.

Moved 'Returns for CStr' section to more appropriate place. Think ultimate goal should be to put it either on '../../concepts/getting-started/type-conversion-functions.md' page, or in a brand-new page documenting just the CStr function.

PLEASE CHECK...
Need to make sure '../../Reference/User-Interface-Help/data-type-summary.md#returns-for-cstr' link will still work if 'Returns for CStr' section is moved as suggested in this commit. If it doesn't, it might be best not to make move.

------------------------------------
THIS COMMIT FOLLOWS CLOSELY ON THE TAIL OF ANOTHER COMMIT TO THIS FILE BY MARKFERN. IT MAY BE BEST TO VIEW ALL MARKFERN'S VERY RECENT COMMITS ON THIS FILE AS A WHOLE.