Fixes & improvements
Various fixes & improvements.

A major problem with previous documentation was that when interfaces were referred to, the reference was to COM interfaces. It was confusing for several reasons. One of which is that the VBE, when it displays Type Library interfaces in the object browser, they are displayed simply as classes. Also, it appears to be common parlance in the VBA developer community to refer to VBA classes that have an 'interface' form, as interfaces when they are also used as such. The example on this page re-affirms this by talking about interface declarations being in a class. Interfaces defined in type libraries are largely hidden from VBA users so it isn't so beneficial to mention them and can be confusing when interfaces also can be used to refer to VBA class-module code.

This commit is best used in relation to the glossary commit at: https://github.com/MarkFern/VBA-Docs/commit/2dca57b54c4b4c90a3ce8f1a0e7f35198443e8e3 . I moved the 'interface' definition from this page into the glossary.