Fixes & improvements; esp. re. ambiguity
Various fixes & improvements.

A major area of concern with previous version was that class implementation could refer to implementation in the class module containing the 'Implements' statement, as well as the implementation in the class module referenced in a use of the 'Implements' statement.

I already understand various OOP principles but for 'newbies', confusion can arise when talking about 'implementing classes' when what is really meant is 'implementing the interfaces represented by classes'.

I've tried to make docs more 'idiot proof'. Now, the term 'interface' also refers to 'interfaces automatically derived from classes'. This, I believe, improves the clarity of the docs a great deal.