OLE VARENUM constant names
Improvements.

Changes mean that information on this page is significantly and appropriately more than corresponding information on the VarType-function page. If information were the same, there would be little point in this page existing.

VARENUM information is useful for understanding Variants and the relationship between COM & VBA. It is also useful when developing solutions involving more than one programming language that rely on the COM technology.