Fixes+improvements; esp. for vbError
- 'Coerced' hyperlink.
- More table footnotes.
- italicised & super-scripted text that isn't so important.
 - vbError also corresponds to missing optional Variant arguments.
- A small number of other small fixes and other small improvements.

I'm hoping that the table cells for the footnotes will be ultimately top-aligned. My preview is showing it as having a centre vertical alignment but I think once the '.md' file for this page has been converted to a proper web-page, the alignment will become a top alignment. This seems to be the case with other VBA doc. web-pages.