Various changes (esp. re. vbError)
- More table footnotes.
- Various texts that don't have central importance, have been italicised and super-scripted.
 - vbError also corresponds to missing optional Variant arguments.

I'm hoping that the table cells for the footnotes will be ultimately top-aligned. My preview is showing it as having a centre vertical alignment but I think once the '.md' file for this page has been converted to a proper web-page, the alignment will become a top alignment. This seems to be the case with other VBA doc. web-pages.