Confusion re. COM object, VBA object, & interface
It can be confusing since the term object refers both to a COM object with all of the COM object's interfaces, as well as to a COM object with one of its interfaces selected (such as in VBA object references).

Have tried to reduce the risk of any such confusion arising. Probably will make more changes for this later.