Confusion re. COM object, VBA object, & interface
It can be confusing since the term object refers both to i) a COM object with all of the COM object's interfaces, as well as to ii) a COM object with one of its interfaces selected (such as in VBA object references). In respect of ii), two objects may be considered different objects even though they refer to the same COM object.

Have tried to reduce the risk of any such confusion arising. Probably will make more changes for this later.